refactor pool card to be reusable
